requirement.txt
Django==4.2
djangorestframework==3.14.0
django-ckeditor==6.6.0
googletrans==4.0.0-rc1
redis==4.5.5
django-redis==5.2.0
pytest==7.3.1
pytest-django==4.5.2

